Fixed unsigned/signed comparison warning

+++ src/priority-bp/impl/Label.h
@@ -1,19 +1,19 @@
 //
 // Copyright 2010, Darren Lafreniere
 // <http://www.lafarren.com/image-completer/>
-// 
+//
 // This file is part of lafarren.com's Image Completer.
-// 
+//
 // Image Completer is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
-// 
+//
 // Image Completer is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
 // along with Image Completer, named License.txt. If not, see
 // <http://www.gnu.org/licenses/>.
@@ -60,7 +60,7 @@ namespace PriorityBp
 
 		const Label& operator[](int i) const;
 		int size() const;
-		
+
 		// Scalable interface
 		virtual void ScaleUp();
 		virtual void ScaleDown();
@@ -74,7 +74,7 @@ namespace PriorityBp
 
 			inline const Label& operator[](int i) const
 			{
-				wxASSERT(i < sizeof(m_labels) / sizeof(m_labels[0]));
+				wxASSERT(static_cast<unsigned int>(i) < sizeof(m_labels) / sizeof(m_labels[0]));
 				return m_labels[i];
 			}
 
@@ -118,7 +118,7 @@ namespace PriorityBp
 			const int m_height;
 			const int m_dataNumUints;
 		};
-		
+
 		struct Resolution
 		{
 			Resolution(const Settings& settings, const Image& inputImage, const MaskLod& mask);
